Refactor dto and entity
