--- backend/src/main/java/com/sprawler/external/myinfo/dto/person/encrypted/EncryptedPersonApiResponse.java
@@ -1,4 +1,0 @@
-package com.sprawler.external.myinfo.dto.person.encrypted;
-
-public class EncryptedPersonApiResponse {
-}
