+++ backend/src/main/java/com/sprawler/external/myinfo/dto/request/TokenRequestDTO.java
@@ -1,0 +1,7 @@
+package com.sprawler.external.myinfo.dto.request;
+
+public record TokenRequestDTO(
+        String code,
+        String verifier
+) {
+}
