+++ backend/src/main/java/com/sprawler/external/myinfo/entity/person/encrypted/EncryptedPersonApiResponse.java
@@ -1,0 +1,4 @@
+package com.sprawler.external.myinfo.entity.person.encrypted;
+
+public class EncryptedPersonApiResponse {
+}
